Move Elasticsearch.Document definitions to a separate directory
Remove Indexer.update_mappings?
Use git diff to detect index-breaking changes in deploy script
