+++ lib/meadow/data/schemas/file_set.ex
@@ -7,7 +7,6 @@ defmodule Meadow.Data.Schemas.FileSet do
   alias Meadow.Data.Schemas.FileSetMetadata
   alias Meadow.Data.Schemas.Work
   alias Meadow.Data.Types
-  alias Meadow.Utils.Exif, as: ExifUtil
 
   import Ecto.Changeset
   import EctoRanked
@@ -80,30 +79,5 @@ defmodule Meadow.Data.Schemas.FileSet do
     |> prepare_embed(:metadata)
     |> cast_embed(:metadata)
     |> set_rank(scope: [:work_id, :role])
-  end
-
-  defimpl Elasticsearch.Document, for: Meadow.Data.Schemas.FileSet do
-    def id(file_set), do: file_set.id
-    def routing(_), do: false
-
-    def encode(file_set) do
-      %{
-        createDate: file_set.inserted_at,
-        description: file_set.metadata.description,
-        label: file_set.metadata.label,
-        mime_type: file_set.metadata.mime_type,
-        model: %{application: "Meadow", name: "FileSet"},
-        modifiedDate: file_set.updated_at,
-        role: format(file_set.role),
-        visibility: format(file_set.work.visibility),
-        workId: file_set.work.id,
-        exif: ExifUtil.index(file_set.metadata.exif)
-      }
-    end
-
-    defp format(%{id: id, name: name}), do: %{id: id, name: name}
-    defp format(%{id: id, title: title}), do: %{id: id, title: title}
-    defp format(%{id: _id, label: _label, scheme: _scheme} = field), do: field
-    defp format(_), do: %{}
   end
 end
